Project re-organized. Folders added and files moved
