+++ src/App.jsx
@@ -1,7 +1,7 @@
 import './App.css'
 import { useState } from 'react';
-import TodoForm from './TodoForm'
-import TodoList from './TodoList'
+import TodoForm from './features/TodoForm'
+import TodoList from './features/TodoList/TodoList'
 
 function App() {
   const [todoList, setTodoList] = useState([]);
@@ -13,7 +13,7 @@ function App() {
   function completeTodo(id) {
     const updatedTodos = todoList.map((todo) => {
       if (todo.id === id) {
-        let value = !todo.isCompleted; // it's prepared for a future uncheck function
+        let value = true;
         return { ...todo, isCompleted: value };
       }
       return todo;
